refactor(#96): useMyChannelsQuery를 useServerApi.ts로 이동
